Brings back webpack.common
