+++ webpack.common.ts
@@ -1,29 +1,34 @@
+import type { Configuration } from "webpack";
+
 const webpack = require("webpack");
 const path = require("path");
 
-module.exports = {
+const config: Configuration = {
   entry: {
-    app: [path.resolve(__dirname, 'src/game.ts')],
+    app: [
+      path.resolve(__dirname, 'src/game.ts')
+    ],
     vendor: ['phaser']
   },
   output: {
     pathinfo: true,
     path: path.resolve(__dirname, 'dist'),
-    publicPath: './dist/',
+    publicPath: '/',
     filename: 'bundle.js'
   },
   module: {
     rules: [
       {
         test: /\.ts$/,
-        use: ['babel-loader'],
-        include: path.join(__dirname, 'src')
+        loaders: ['babel-loader'],
+        include: path.join(__dirname, 'src'),
       },
-      // {
-      //   test: /\.ts$/,
-      //   exclude: /node_module/,
-      //   use: 'ts-loader'
-      // }
+      { // https://v4.webpack.js.org/guides/asset-management/
+        test: /\.(png|svg|jpg|gif)$/,
+        use: [
+          'file-loader',
+        ],
+      }
     ]
   },
   node: {
@@ -31,17 +36,19 @@ module.exports = {
     net: 'empty',
     tls: 'empty'
   },
-  // target: 'node',
   optimization: {
     splitChunks: {
       name: 'vendor',
       chunks: 'all'
     }
   },
+  // // target: 'node',
   plugins: [
     new webpack.DefinePlugin({
       CANVAS_RENDERER: JSON.stringify(true),
       WEBGL_RENDERER: JSON.stringify(true)
     })
   ]
-}
+};
+
+export default config;
